add postprocessing script

+++ bins/SAGA/osk_postprocess.sh
@@ -23,18 +23,30 @@ source /home/avollrath/github/OpenSARKit/OpenSARKit_source.bash
 	
 # 	0.1 Check for right usage
 if [ "$#" != "1" ]; then
-  echo -e "----------------------------------------------------------------------------------------------"
-  echo -e "Usage: osk_mosaic_tracks </path/to/mosaic/sat/tracks>"
-  echo -e "This script mosaics primarily merged Satellite Tracks (i.e. output of osk_SENSOR_merge_path)"
-  echo -e "----------------------------------------------------------------------------------------------"
+  echo -e "------------------------------------------------------------------------------------------------------------------------"
+  echo -e "Usage: osk_postprocess </path/to/mosaic/sat/tracks>"
+  echo -e "------------------------------------------------------------------------------------------------------------------------"
+  echo -e "This script: mosaics primarily merged Satellite Tracks (i.e. output of osk_SENSOR_merge_path)"
+  echo -e "	- mosaics primarily merged Satellite Tracks (i.e. output of osk_SENSOR_merge_path) [SAGA GIS]"
+  echo -e "	- filters the Gamma0 channels with a multi-directional filter (Lee et al. 1998) [SAGA GIS]"
+  echo -e "	- calculates second order statistics for Texture measurements (i.e. GLCM Mean, GLCM Variance) [Sentinel Toolbox]"
+  echo -e "	- outputs all files with sam extent and resolution in GeoTiff format ready for RSGISLIB segmentation [gdal]"
+  echo -e "------------------------------------------------------------------------------------------------------------------------"
   echo -e "The path will be your Project folder!"
   exit 1
 else
   cd $1
   export PROC_DIR=`pwd`
-  echo -e "This script mosaics primarily merged Satellite Tracks (i.e. output of osk_SENSOR_merge_path)"
+  echo -e "------------------------------------------------------------------------------------------------------------------------"
+  echo -e "This script: mosaics primarily merged Satellite Tracks (i.e. output of osk_SENSOR_merge_path)"
+  echo -e "	- mosaics primarily merged Satellite Tracks (i.e. output of osk_SENSOR_merge_path) [SAGA GIS]"
+  echo -e "	- filters the Gamma0 channels with a multi-directional filter (Lee et al. 1998) [SAGA GIS]"
+  echo -e "	- calculates second order statistics for Texture measurements (i.e. GLCM Mean, GLCM Variance) [Sentinel Toolbox]"
+  echo -e "	- outputs all files with sam extent and resolution in GeoTiff format ready for RSGISLIB segmentation [gdal]"
+  echo -e "------------------------------------------------------------------------------------------------------------------------"
   echo "Processing folder: ${PROC_DIR}"
-  echo "Output folder: ${PROC_DIR}/FINAL_MOSAIC"
+  echo "Output folder: ${PROC_DIR}/../FINAL_MOSAIC"
+  echo -e "------------------------------------------------------------------------------------------------------------------------"
 fi
 
 # set up paths
@@ -48,7 +60,7 @@ mkdir -p $TMP_DIR
 # reduce range from -30 to 5db for HH channel and mosaicing
 #------------------------------------------------------------
 
-echo "reduce range from -30 to 5db for HH channel"
+echo "Reducing range from -30 to 5db for HH channel"
 for line in `ls -1 *Gamma0_HH_db.sdat`;do 
 	gdal_calc.py -A $line --outfile ${TMP_DIR}/tmp1.tif --calc="A*(A>-30)" --NoDataValue=-99999
 	gdal_calc.py -A ${TMP_DIR}/tmp1.tif  --outfile ${TMP_DIR}/tmp2.tif --calc="A*(A<5)" --NoDataValue=-99999
@@ -56,12 +68,20 @@ for line in `ls -1 *Gamma0_HH_db.sdat`;do
 	rm -f ${TMP_DIR}/tmp*
 done 
 
-echo "mosaicing the HH channel"
-LIST_GAMMA_HH=`ls -1 end*Gamma0_HH_db.sgrd | tr '\n' ';'`
-saga_cmd grid_tools 3 -GRIDS:${LIST_GAMMA_HH} -TYPE:7 -OVERLAP:6 -BLEND_DIST:10	 -MATCH:1 -TARGET_OUT_GRID:${TMP_DIR}/GAMMA0_HH_db_mosaic.sgrd
-rm -rf ${TMP_DIR}/end*
+echo "Mosaicing the HH channel"
+if [[ `ls -1 ${TMP_DIR}/end*Gamma0_HH_db.sgrd | wc -l` == 1 ]];then
+	mv ${TMP_DIR}/end*sgrd ${TMP_DIR}/GAMMA0_HH_db_mosaic.sgrd
+	mv ${TMP_DIR}/end*mgrd ${TMP_DIR}/GAMMA0_HH_db_mosaic.mgrd
+	mv ${TMP_DIR}/end*sdat ${TMP_DIR}/GAMMA0_HH_db_mosaic.sdat
+	mv ${TMP_DIR}/end*prj ${TMP_DIR}/GAMMA0_HH_db_mosaic.prj
+else
+	LIST_GAMMA_HH=`ls -1 ${TMP_DIR}/end*Gamma0_HH_db.sgrd | tr '\n' ';'`
+	saga_cmd grid_tools 3 -GRIDS:${LIST_GAMMA_HH} -TYPE:7 -OVERLAP:6 -BLEND_DIST:10	 -MATCH:1 -TARGET_OUT_GRID:${TMP_DIR}/GAMMA0_HH_db_mosaic.sgrd
+	rm -rf ${TMP_DIR}/end*
+fi 
 
-echo "reduce range from -30 to 5db for HV channel"
+
+echo "Reducing range from -30 to 5db for HV channel"
 for line in `ls -1 *Gamma0_HV_db.sdat`;do 
 	gdal_calc.py -A $line --outfile ${TMP_DIR}/tmp1.tif --calc="A*(A>-30)" --NoDataValue=-99999
 	gdal_calc.py -A ${TMP_DIR}/tmp1.tif  --outfile ${TMP_DIR}/tmp2.tif --calc="A*(A<5)" --NoDataValue=-99999
@@ -69,11 +89,18 @@ for line in `ls -1 *Gamma0_HV_db.sdat`;do
 	rm -f ${TMP_DIR}/tmp*
 done 
 
-echo "mosaicing the HV channel"
-LIST_GAMMA_HV=`ls end*Gamma0_HV_db.sgrd | tr '\n' ';'`
-saga_cmd grid_tools 3 -GRIDS:${LIST_GAMMA_HV} -TYPE:7 -OVERLAP:6 -BLEND_DIST:10 -MATCH:1 -TARGET_OUT_GRID:${TMP_DIR}/GAMMA0_HV_db_mosaic.sgrd
-rm -rf ${TMP_DIR}/end*
+echo "Mosaicing the HV channel"
+if [[ `ls -1 ${TMP_DIR}/end*Gamma0_HV_db.sgrd | wc -l` == 1 ]];then
+	mv ${TMP_DIR}/end*sgrd ${TMP_DIR}/GAMMA0_HV_db_mosaic.sgrd
+	mv ${TMP_DIR}/end*mgrd ${TMP_DIR}/GAMMA0_HV_db_mosaic.mgrd
+	mv ${TMP_DIR}/end*sdat ${TMP_DIR}/GAMMA0_HV_db_mosaic.sdat
+	mv ${TMP_DIR}/end*prj ${TMP_DIR}/GAMMA0_HV_db_mosaic.prj
 
+else
+	LIST_GAMMA_HV=`ls ${TMP_DIR}/end*Gamma0_HV_db.sgrd | tr '\n' ';'`
+	saga_cmd grid_tools 3 -GRIDS:${LIST_GAMMA_HV} -TYPE:7 -OVERLAP:6 -BLEND_DIST:10 -MATCH:1 -TARGET_OUT_GRID:${TMP_DIR}/GAMMA0_HV_db_mosaic.sgrd
+	rm -rf ${TMP_DIR}/end*
+fi
 
 echo "--------------------------------------------------------------------------------------------------"
 echo " Calculate second order statistic Texture measurements (GLCMMean, GLCMVariance) for HH channel"
@@ -132,29 +159,29 @@ echo "------------------------------------------------------------"
 echo " Multi-directional filtering of the Gamma0 HH channel"
 echo "------------------------------------------------------------"
 gdalwarp -of SAGA -srcnodata 0 -dstnodata -99999 ${TMP_DIR}/TEXTURE_HH.data/band_1.img ${TMP_DIR}/TEXTURE_HH.data/band_1_saga.sdat
-saga_cmd -f=r -c=8 grid_filter 3 -INPUT:${TMP_DIR}/band_1_saga.sgrd -RESULT:${TMP_DIR}/GAMMA0_HH_db_mosaic_filtered.sgrd -NOISE_ABS:5 -NOISE_REL:3 -METHOD:1
+saga_cmd -f=r grid_filter 3 -INPUT:${TMP_DIR}/TEXTURE_HH.data/band_1_saga.sgrd -RESULT:${TMP_DIR}/GAMMA0_HH_db_mosaic_filtered.sgrd -NOISE_ABS:5 -NOISE_REL:3 -METHOD:1
 
 echo "------------------------------------------------------------"
 echo " Multi-directional filtering of the Gamma0 HV channel"
 echo "------------------------------------------------------------"
 gdalwarp -of SAGA -srcnodata 0 -dstnodata -99999 ${TMP_DIR}/TEXTURE_HV.data/band_1.img ${TMP_DIR}/TEXTURE_HV.data/band_1_saga.sdat
-saga_cmd -f=r -c=8 grid_filter 3 -INPUT:${TMP_DIR}/TEXTURE_HV.data/band_1_saga.sdat -RESULT:${TMP_DIR}/GAMMA0_HV_db_mosaic_filtered.sgrd -NOISE_ABS:5 -NOISE_REL:3 -METHOD:1
+saga_cmd -f=r grid_filter 3 -INPUT:${TMP_DIR}/TEXTURE_HV.data/band_1_saga.sdat -RESULT:${TMP_DIR}/GAMMA0_HV_db_mosaic_filtered.sgrd -NOISE_ABS:5 -NOISE_REL:3 -METHOD:1
 
 echo "------------------------------------------------------------"
 echo " Calculate HH/HV ratio"
 echo "------------------------------------------------------------"
 #saga_cmd grid_calculus 1 -GRIDS:${FINAL}/GAMMA0_HH_db_mosaic.sdat -XGRIDS:${FINAL}/GAMMA0_HV_db_mosaic.sdat -RESULT:${FINAL}/HH_HV_db_mosaic.sdat -FORMULA:"a / b"
-saga_cmd -f=r -c=8 grid_calculus 1 -GRIDS:${TMP_DIR/GAMMA0_HH_db_mosaic_filtered.sdat -XGRIDS:${TMP_DIR}/GAMMA0_HV_db_mosaic_filtered.sdat -RESULT:${TMP_DIR}/HH_HV_db_mosaic_filtered.sdat -FORMULA:"a / b"
+saga_cmd -f=r grid_calculus 1 -GRIDS:${TMP_DIR}/GAMMA0_HH_db_mosaic_filtered.sdat -XGRIDS:${TMP_DIR}/GAMMA0_HV_db_mosaic_filtered.sdat -RESULT:${TMP_DIR}/HH_HV_db_mosaic_filtered.sdat -FORMULA:"a / b"
 
-gdalwarp -of GTiff -srcnodata -99999 -dstnodata 0 ${TMP_DIR/GAMMA0_HH_db_mosaic_filtered.sdat ${FINAL}/Gamma0_HH.tif
-gdalwarp -of GTiff -srcnodata -99999 -dstnodata 0 ${TMP_DIR/GAMMA0_HV_db_mosaic_filtered.sdat ${FINAL}/Gamma0_HV.tif
+gdalwarp -of GTiff -srcnodata -99999 -dstnodata 0 ${TMP_DIR}/GAMMA0_HH_db_mosaic_filtered.sdat ${FINAL}/Gamma0_HH.tif
+gdalwarp -of GTiff -srcnodata -99999 -dstnodata 0 ${TMP_DIR}/GAMMA0_HV_db_mosaic_filtered.sdat ${FINAL}/Gamma0_HV.tif
 gdalwarp -of GTiff -srcnodata -99999 -dstnodata 0 ${TMP_DIR}/HH_HV_db_mosaic_filtered.sdat ${FINAL}/HH_HV_ratio.tif
 gdalwarp -of GTiff -srcnodata 0 -dstnodata 0 ${TMP_DIR}/TEXTURE_HH.data/GLCMMean.img ${FINAL}/GLCMMean_HH.tif
 gdalwarp -of GTiff -srcnodata 0 -dstnodata 0 ${TMP_DIR}/TEXTURE_HH.data/GLCMVariance.img ${FINAL}/GLCMVariance_HH.tif
 gdalwarp -of GTiff -srcnodata 0 -dstnodata 0 ${TMP_DIR}/TEXTURE_HV.data/GLCMMean.img ${FINAL}/GLCMMean_HV.tif
 gdalwarp -of GTiff -srcnodata 0 -dstnodata 0 ${TMP_DIR}/TEXTURE_HV.data/GLCMVariance.img ${FINAL}/GLCMVariance_HV.tif
 
-rm -rf ${TMP_DIR}
+#rm -rf ${TMP_DIR}
 
 echo "----------------------------------"
 echo " Processing succesfully finished"
